requirements.txt
# Telegram Bot Framework
aiogram>=3.22.0

# HTTP Client (совместимая версия с aiogram 3.22.0)
aiohttp>=3.12.0

# Database
aiosqlite==0.18.0

# Environment Variables
python-dotenv==1.0.0

# Scheduling (используется в проекте, но можно удалить если не нужно)
apscheduler==3.10.1

# WebSocket для арбитражного бота
websockets==12.0

# Опционально: uvloop для лучшей производительности
uvloop==0.19.0; sys_platform != 'win32'

# Plotting for stats
matplotlib>=3.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
